fix(flutter): Relocate home_page.dart
